convert to int for min_pixels

+++ openrlhf/trainer/ppo_utils/data_processor.py
@@ -79,8 +79,8 @@ class BaseDataProcessor(ABC):
 
 def add_pixel_bounds(messages):
     # 默认的像素范围
-    DEFAULT_MIN_PIXELS = os.getenv("MIN_PIXELS", 4 * 28 * 28)
-    DEFAULT_MAX_PIXELS = os.getenv("MAX_PIXELS", 640 * 28 * 28)
+    DEFAULT_MIN_PIXELS = int(os.getenv("MIN_PIXELS", 4 * 28 * 28))
+    DEFAULT_MAX_PIXELS = int(os.getenv("MAX_PIXELS", 640 * 28 * 28))
 
     def process_content(content):
         if isinstance(content, list):
